+++ scripts/_helpers/data.ts
@@ -1,0 +1,3 @@
+export const PROXY_ADMIN_ADDRESS: string = "";
+export const DLOGOS_IMPLEMENTATION_ADDRESS: string = "";
+export const DLOGOS_INSTANCE_ADDRESS: string = "";